feat: added composed string and URL PrototypeString types
