+++ src/main/kotlin/org/coralprotocol/coralserver/agent/registry/RegistryAgent.kt
@@ -1,26 +1,13 @@
-@file:OptIn(ExperimentalSerializationApi::class, InternalSerializationApi::class)
-
 package org.coralprotocol.coralserver.agent.registry
 
-import io.github.z4kn4fein.semver.Version
-import io.github.z4kn4fein.semver.VersionFormatException
-import kotlinx.serialization.*
-import me.saket.bytesize.BinaryByteSize
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
 import me.saket.bytesize.kibibytes
 import me.saket.bytesize.mebibytes
-import org.bitcoinj.core.AddressFormatException
-import org.bitcoinj.core.Base58
 import org.coralprotocol.coralserver.agent.registry.option.AgentOption
 import org.coralprotocol.coralserver.agent.registry.option.defaultAsValue
 import org.coralprotocol.coralserver.agent.runtime.LocalAgentRuntimes
-import org.coralprotocol.coralserver.agent.runtime.PrototypeRuntime
 import org.coralprotocol.coralserver.agent.runtime.RuntimeId
-import org.coralprotocol.coralserver.agent.runtime.prototype.PrototypeApiUrl
-import org.coralprotocol.coralserver.agent.runtime.prototype.PrototypeString
-import org.coralprotocol.coralserver.agent.runtime.prototype.PrototypeToolServer
-import org.coralprotocol.coralserver.agent.runtime.prototype.PrototypeToolServerAuth
-import java.net.URI
-import java.net.URISyntaxException
 import java.nio.file.Path
 
 /**
@@ -32,68 +19,6 @@ const val MINIMUM_SUPPORTED_AGENT_EDITION = 3
  * The maximum (and current) supported agent edition.
  */
 const val MAXIMUM_SUPPORTED_AGENT_VERSION = 4
-
-// [agent]
-val AGENT_NAME_LENGTH = 1..32
-val AGENT_NAME_PATTERN = "^[a-z0-9]([a-z0-9]*(-[a-z0-9]+)*)?$".toRegex()
-val AGENT_VERSION_LENGTH = 1..24
-val AGENT_DESCRIPTION_LENGTH = 1..1024
-val AGENT_SUMMARY_LENGTH = 1..256
-val AGENT_README_MAX_SIZE = 1..4096
-val AGENT_LICENSE_TEXT_MAX_SIZE = 2.mebibytes
-const val AGENT_KEYWORDS_MAX_ENTRIES = 256
-val AGENT_KEYWORDS_LENGTH = 1..64
-
-// [agent.links]
-const val AGENT_LINKS_MAX_ENTRIES = 16
-val AGENT_LINKS_NAME_LENGTH = 1..32
-val AGENT_LINKS_NAME_PATTERN = "^[a-zA-Z][a-zA-Z_\\-0-9]*$".toRegex()
-val AGENT_LINK_VALUE_LENGTH = 1..256
-
-// [runtimes.docker]
-val AGENT_DOCKER_IMAGE_LENGTH = 1..512
-val AGENT_DOCKER_COMMAND_ENTRIES = 0..1024
-val AGENT_DOCKER_COMMAND_MAX_SIZE = 2.kibibytes
-
-// [runtimes.executable]
-val AGENT_EXECUTABLE_PATH_LENGTH = 1..4096
-val AGENT_EXECUTABLE_ARGUMENTS_ENTRIES = 0..1024
-val AGENT_EXECUTABLE_ARGUMENTS_SIZE = 2.kibibytes
-
-// [runtimes.prototype]
-val AGENT_PROTOTYPE_MODEL_NAME_LENGTH = 1..128
-val AGENT_PROTOTYPE_MODEL_KEY_LENGTH = 1..1024
-val AGENT_PROTOTYPE_MODEL_API_URL_LENGTH = 1..256
-val AGENT_PROTOTYPE_MCP_TOOL_SERVER_URL_LENGTH = 1..256
-val AGENT_PROTOTYPE_MCP_AUTH_BEARER_LENGTH = 1..1024
-const val AGENT_PROTOTYPE_MCP_AUTH_TRANSFORMATION_MAX_ENTRIES = 8
-val AGENT_PROTOTYPE_MCP_AUTH_TRANSFORMATION_PATTERN_LENGTH = 1..32
-val AGENT_PROTOTYPE_MCP_AUTH_TRANSFORMATION_REPLACEMENT_LENGTH = 1..1024
-val AGENT_PROTOTYPE_PROMPT_SYSTEM_BASE_SIZE = 6.kibibytes
-val AGENT_PROTOTYPE_PROMPT_SYSTEM_EXTRA_SIZE = 3.kibibytes
-val AGENT_PROTOTYPE_PROMPT_LOOP_INITIAL_BASE_SIZE = 6.kibibytes
-val AGENT_PROTOTYPE_PROMPT_LOOP_INITIAL_EXTRA_SIZE = 3.kibibytes
-val AGENT_PROTOTYPE_PROMPT_LOOP_FOLLOWUP_SIZE = 4.kibibytes
-
-// [options]
-const val AGENT_OPTION_MAX_ENTRIES = 512
-val AGENT_OPTION_NAME_LENGTH = 1..256
-val AGENT_OPTION_NAME_PATTERN = "^[a-zA-Z_][a-zA-Z_$0-9]*$".toRegex()
-val AGENT_OPTION_DEFAULTS_MAX_SIZE = 6.mebibytes
-val AGENT_OPTION_DISPLAY_LABEL_LENGTH = 1..64
-val AGENT_OPTION_DISPLAY_DESCRIPTION_LENGTH = 1..1024
-val AGENT_OPTION_DISPLAY_GROUP_LENGTH = 1..64
-
-// [marketplace.pricing]
-val AGENT_MARKETPLACE_PRICING_DESCRIPTION_LENGTH = 1..256
-const val AGENT_MARKETPLACE_PRICING_MIN_MIN = 0.00
-const val AGENT_MARKETPLACE_PRICING_MIN_MAX = 20.00
-
-// [marketplace.identities.erc8004]
-const val AGENT_MARKETPLACE_ERC8004_ENDPOINTS_MAX_ENTRIES = 32
-val AGENT_MARKETPLACE_ERC8004_ENDPOINTS_NAME_LENGTH = 1..32
-val AGENT_MARKETPLACE_ERC8004_ENDPOINTS_NAME_PATTERN = "^[a-zA-Z][a-zA-Z_\\-0-9]*$".toRegex()
-val AGENT_MARKETPLACE_ERC8004_ENDPOINTS_ENDPOINT_LENGTH = 1..256
 
 @Serializable
 data class RegistryAgent(
@@ -151,409 +76,6 @@ data class RegistryAgent(
     @Transient
     val requiredOptions = options
         .filterValues { it.required }
-
-    private fun validateStringLength(name: String, string: String, range: IntRange) {
-        if (range.first > 0 && string.isEmpty())
-            throw RegistryException("\"$name\" must not be empty")
-
-        if (string.length < range.first)
-            throw RegistryException("\"$name\" must be at least ${range.first} characters long, was ${string.length}")
-
-        if (string.length > range.last)
-            throw RegistryException("\"$name\" must be at most ${range.last} characters long, was ${string.length}")
-    }
-
-    private fun validateStringList(
-        name: String,
-        list: List<String>,
-        listRange: IntRange,
-        maxTotalSize: BinaryByteSize
-    ) {
-        if (listRange.first > 0 && listRange.isEmpty())
-            throw RegistryException("\"$name\" must not be empty")
-
-        if (list.size < listRange.first)
-            throw RegistryException("\"$name\" must have at least ${listRange.first} entries, has ${list.size}")
-
-        if (list.size > listRange.last)
-            throw RegistryException("\"$name\" must have at most ${listRange.last} entries, has ${list.size}")
-
-        val size = BinaryByteSize(list.sumOf { it.toByteArray().size })
-        if (size > maxTotalSize)
-            throw RegistryException("total size for \"$name\" must be at most $maxTotalSize, was $size")
-    }
-
-    private fun validateUri(
-        name: String,
-        value: String,
-        lengthRange: IntRange,
-        vararg allowedSchemes: String
-    ) {
-        validateStringLength(name, value, lengthRange)
-
-        try {
-            val uri = URI(value)
-            if (uri.scheme !in allowedSchemes) {
-                throw RegistryException(
-                    "\"$name\" has invalid schema \"${uri.scheme}\", must be: ${
-                        allowedSchemes.joinToString(
-                            ", "
-                        )
-                    }"
-                )
-            }
-        } catch (e: URISyntaxException) {
-            throw RegistryException("\"$name\" is not a valid URL: ${e.message}")
-        }
-    }
-
-    private fun PrototypeString.validatePrototypeString(name: String, inlineLength: IntRange) {
-        when (this) {
-            is PrototypeString.Inline -> validateStringLength(name, value, inlineLength)
-            is PrototypeString.Option -> {
-                val option = options[this.name]
-                    ?: throw RegistryException("\"$name\" references option \"${this.name}\" which is not defined")
-
-                if (option !is AgentOption.String)
-                    throw RegistryException("\"$name\" references option \"${this.name}\" which must be a string type, was ${option::class.serializer().descriptor.serialName}")
-            }
-        }
-    }
-
-    private fun PrototypeString.validatePrototypeString(name: String, maxSize: BinaryByteSize) {
-        when (this) {
-            is PrototypeString.Inline -> {
-                val byteSize = BinaryByteSize(value.toByteArray().size)
-                if (byteSize > maxSize)
-                    throw RegistryException("\"$name\" cannot exceed ${maxSize}, was $byteSize")
-            }
-
-            is PrototypeString.Option -> validatePrototypeString(name, 0..0)
-        }
-    }
-
-    private fun validateName() {
-        validateStringLength("agent.name", name, AGENT_NAME_LENGTH)
-
-        if (!name.matches(AGENT_NAME_PATTERN))
-            throw RegistryException("value for \"agent.name\" ($name) must start with a lowercase alphabetic character and contain only lowercase alphanumeric characters or '-'")
-    }
-
-    private fun validateVersion() {
-        validateStringLength("agent.version", version, AGENT_VERSION_LENGTH)
-
-        try {
-            Version.parse(version)
-        } catch (e: VersionFormatException) {
-            throw RegistryException("invalid version provided for \"agent.version\": ${e.message}")
-        }
-    }
-
-    private fun validateOptionalAgentInfo() {
-        validateStringLength("agent.description", description, AGENT_DESCRIPTION_LENGTH)
-        validateStringLength("agent.summary", summary, AGENT_SUMMARY_LENGTH)
-        validateStringLength("agent.readme", readme, AGENT_README_MAX_SIZE)
-
-        when (license) {
-            is RegistryAgentLicense.Spdx -> {
-                // TODO
-            }
-
-            is RegistryAgentLicense.Text -> {
-                val size = BinaryByteSize(license.text.toByteArray().size)
-                if (size > AGENT_LICENSE_TEXT_MAX_SIZE) {
-                    throw RegistryException("agent license text size count cannot exceed $AGENT_LICENSE_TEXT_MAX_SIZE, was $size")
-                }
-            }
-        }
-
-        if (links.size > AGENT_LINKS_MAX_ENTRIES)
-            throw RegistryException("agent link count cannot exceed $AGENT_LINKS_MAX_ENTRIES, was ${links.size}")
-
-        if (keywords.size > AGENT_KEYWORDS_MAX_ENTRIES)
-            throw RegistryException("number of agent keywords cannot exceed $AGENT_KEYWORDS_MAX_ENTRIES, was ${keywords.size}")
-
-        keywords.forEachIndexed { index, keyword ->
-            validateStringLength(
-                "agent.keywords[$index]",
-                keyword,
-                AGENT_KEYWORDS_LENGTH
-            )
-        }
-
-        for ((name, link) in links) {
-            validateStringLength("agent.links[\"$name\"] (key)", name, AGENT_LINKS_NAME_LENGTH)
-
-            if (!name.matches(AGENT_LINKS_NAME_PATTERN))
-                throw RegistryException("agent link \"$name\" is not valid.  Agent link names must start with an alphabetic character and contain only alphanumeric characters or underscores")
-
-            validateUri(
-                "agent.links[\"$name\"] (value)",
-                link,
-                AGENT_LINK_VALUE_LENGTH,
-                "https", "mailto", "tel"
-            )
-        }
-    }
-
-    private fun validateRuntimes() {
-        if (runtimes.functionRuntime == null && runtimes.dockerRuntime == null && runtimes.executableRuntime == null && runtimes.prototypeRuntime == null)
-            throw RegistryException("Must have at least one defined runtime")
-
-        val docker = runtimes.dockerRuntime
-        if (docker != null) {
-            validateStringLength("runtimes.docker.image", docker.image, AGENT_DOCKER_IMAGE_LENGTH)
-
-            if (docker.command != null) {
-                validateStringList(
-                    "runtimes.docker.command",
-                    docker.command,
-                    AGENT_DOCKER_COMMAND_ENTRIES,
-                    AGENT_DOCKER_COMMAND_MAX_SIZE
-                )
-            }
-        }
-
-        val executable = runtimes.executableRuntime
-        if (executable != null) {
-            validateStringLength("runtimes.executable.path", executable.path, AGENT_EXECUTABLE_PATH_LENGTH)
-
-            validateStringList(
-                "runtimes.executable.arguments",
-                executable.arguments,
-                AGENT_EXECUTABLE_ARGUMENTS_ENTRIES,
-                AGENT_EXECUTABLE_ARGUMENTS_SIZE
-            )
-        }
-
-        if (runtimes.prototypeRuntime != null)
-            validatePrototypeRuntime(runtimes.prototypeRuntime)
-    }
-
-    private fun validatePrototypeRuntime(runtime: PrototypeRuntime) {
-        if (runtime.iterationCount <= 0)
-            throw RegistryException("\"runtimes.prototype.iterations\" must be at least 1")
-
-        if (runtime.iterationDelay < 0)
-            throw RegistryException("\"runtimes.prototype.delay\" cannot be negative")
-
-        val model = runtime.modelProvider
-        model.name.validatePrototypeString("runtimes.prototype.model.name", AGENT_PROTOTYPE_MODEL_NAME_LENGTH)
-        model.key.validatePrototypeString("runtimes.prototype.model.key", AGENT_PROTOTYPE_MODEL_KEY_LENGTH)
-
-
-        val customUrl = (model.url as? PrototypeApiUrl.Custom)?.value
-        if (customUrl != null) {
-            validateUri(
-                "runtimes.prototype.model.url.value",
-                customUrl,
-                AGENT_PROTOTYPE_MODEL_API_URL_LENGTH,
-                "https"
-            )
-        }
-
-        runtime.prompts.system.base.validatePrototypeString(
-            "runtimes.prototype.prompts.system.base",
-            AGENT_PROTOTYPE_PROMPT_SYSTEM_BASE_SIZE
-        )
-
-        runtime.prompts.system.extra?.validatePrototypeString(
-            "runtimes.prototype.prompts.system.extra",
-            AGENT_PROTOTYPE_PROMPT_SYSTEM_EXTRA_SIZE
-        )
-
-        runtime.prompts.loop.initial.base.validatePrototypeString(
-            "runtimes.prototype.prompts.loop.initial.base",
-            AGENT_PROTOTYPE_PROMPT_LOOP_INITIAL_BASE_SIZE
-        )
-
-        runtime.prompts.loop.initial.extra?.validatePrototypeString(
-            "runtimes.prototype.prompts.loop.initial.base",
-            AGENT_PROTOTYPE_PROMPT_LOOP_INITIAL_EXTRA_SIZE
-        )
-
-        runtime.prompts.loop.followup.validatePrototypeString(
-            "runtimes.prototype.prompts.loop.followup",
-            AGENT_PROTOTYPE_PROMPT_LOOP_FOLLOWUP_SIZE
-        )
-
-        for ((toolIndex, toolServer) in runtime.toolServers.withIndex()) {
-            val (url, auth) = when (toolServer) {
-                is PrototypeToolServer.McpSse -> Pair(toolServer.url, toolServer.auth)
-                is PrototypeToolServer.McpStreamableHttp -> Pair(toolServer.url, toolServer.auth)
-                else -> throw RegistryException("no tool server validation implemented for ${toolServer::class.serializer().descriptor.serialName}")
-            }
-
-            validateStringLength(
-                "runtimes.prototype.tools[$toolIndex].url",
-                url,
-                AGENT_PROTOTYPE_MCP_TOOL_SERVER_URL_LENGTH
-            )
-
-            when (auth) {
-                is PrototypeToolServerAuth.Bearer -> {
-                    auth.token.validatePrototypeString(
-                        "runtimes.prototype.tools[$toolIndex].auth.token",
-                        AGENT_PROTOTYPE_MCP_AUTH_BEARER_LENGTH
-                    )
-                }
-
-                is PrototypeToolServerAuth.UrlTransformation -> {
-                    if (auth.transformations.size > AGENT_PROTOTYPE_MCP_AUTH_TRANSFORMATION_MAX_ENTRIES)
-                        throw RegistryException("runtimes.prototype.tools[$toolIndex].auth.transformations cannot exceed $AGENT_PROTOTYPE_MCP_AUTH_TRANSFORMATION_MAX_ENTRIES, was ${auth.transformations.size}")
-
-                    auth.transformations.withIndex().forEach { (index, transform) ->
-                        validateStringLength(
-                            "runtimes.prototype.tools[$toolIndex].auth.transformations[$index].pattern",
-                            transform.pattern,
-                            AGENT_PROTOTYPE_MCP_AUTH_TRANSFORMATION_PATTERN_LENGTH
-                        )
-
-                        transform.replacement.validatePrototypeString(
-                            "runtimes.prototype.tools[$toolIndex].auth.transformations[$index].replacement",
-                            AGENT_PROTOTYPE_MCP_AUTH_TRANSFORMATION_REPLACEMENT_LENGTH
-                        )
-                    }
-                }
-
-                PrototypeToolServerAuth.None -> {}
-            }
-        }
-    }
-
-    private fun validateOptions() {
-        if (options.size > AGENT_OPTION_MAX_ENTRIES)
-            throw RegistryException("option count cannot exceed $AGENT_OPTION_MAX_ENTRIES, found ${options.size} defined options")
-
-        var accumulatedDefaultSize = BinaryByteSize(0)
-        for ((name, option) in options) {
-            validateStringLength("options.$name", name, AGENT_OPTION_NAME_LENGTH)
-
-            if (!name.matches(AGENT_OPTION_NAME_PATTERN))
-                throw RegistryException("option name \"$name\" is not valid.  Option names must start with an alphabetic character or underscore and contain only alphanumeric characters or underscores")
-
-            val label = option.display?.label
-            if (label != null)
-                validateStringLength("options.$name.display.label", label, AGENT_OPTION_DISPLAY_LABEL_LENGTH)
-
-            val description = option.display?.description
-            if (description != null) {
-                validateStringLength(
-                    "options.$name.display.description",
-                    description,
-                    AGENT_OPTION_DISPLAY_DESCRIPTION_LENGTH
-                )
-            }
-
-            val group = option.display?.group
-            if (group != null)
-                validateStringLength("options.$name.display.group", group, AGENT_OPTION_DISPLAY_GROUP_LENGTH)
-
-            accumulatedDefaultSize += BinaryByteSize(
-                when (option) {
-                    is AgentOption.Blob -> option.defaultBytes?.size ?: 0
-                    is AgentOption.BlobList -> option.defaultBytes.sumOf { it.size }
-                    is AgentOption.Boolean -> option.default?.let { 1 } ?: 0
-                    is AgentOption.Byte -> option.default?.let { Byte.SIZE_BYTES } ?: 0
-                    is AgentOption.ByteList -> option.default.size
-                    is AgentOption.Double -> option.default?.let { Double.SIZE_BYTES } ?: 0
-                    is AgentOption.DoubleList -> option.default.size * Double.SIZE_BYTES
-                    is AgentOption.Float -> option.default?.let { Float.SIZE_BYTES } ?: 0
-                    is AgentOption.FloatList -> option.default.size * Float.SIZE_BYTES
-                    is AgentOption.Int -> option.default?.let { Int.SIZE_BYTES } ?: 0
-                    is AgentOption.IntList -> option.default.size * Int.SIZE_BYTES
-                    is AgentOption.Long -> option.default?.let { Long.SIZE_BYTES } ?: 0
-                    is AgentOption.LongList -> option.default.size * Long.SIZE_BYTES
-                    is AgentOption.Short -> option.default?.let { Short.SIZE_BYTES } ?: 0
-                    is AgentOption.ShortList -> option.default.size * Short.SIZE_BYTES
-                    is AgentOption.String -> option.default?.toByteArray()?.size ?: 0
-                    is AgentOption.StringList -> option.default.sumOf { it.toByteArray().size }
-                    is AgentOption.UByte -> option.default?.let { UByte.SIZE_BYTES } ?: 0
-                    is AgentOption.UByteList -> option.default.size * UByte.SIZE_BYTES
-                    is AgentOption.UInt -> option.default?.let { UInt.SIZE_BYTES } ?: 0
-                    is AgentOption.UIntList -> option.default.size * UInt.SIZE_BYTES
-                    is AgentOption.ULong -> option.default?.toByteArray()?.size ?: 0
-                    is AgentOption.ULongList -> option.default.sumOf { it.toByteArray().size }
-                    is AgentOption.UShort -> option.default?.let { UShort.SIZE_BYTES } ?: 0
-                    is AgentOption.UShortList -> option.default.size * UShort.SIZE_BYTES
-                })
-        }
-
-        if (accumulatedDefaultSize > AGENT_OPTION_DEFAULTS_MAX_SIZE)
-            throw RegistryException("total size for all default values cannot exceed $AGENT_OPTION_DEFAULTS_MAX_SIZE, was $accumulatedDefaultSize")
-    }
-
-    private fun validateMarketplace() {
-        if (marketplace == null)
-            return
-
-        val pricing = marketplace.pricing
-        if (pricing != null) {
-            validateStringLength(
-                "marketplace.pricing.description",
-                pricing.description,
-                AGENT_MARKETPLACE_PRICING_DESCRIPTION_LENGTH
-            )
-
-            if (pricing.currency != "USD")
-                throw RegistryException("marketplace pricing currency must be USD")
-
-            if (pricing.recommendations.min < AGENT_MARKETPLACE_PRICING_MIN_MIN)
-                throw RegistryException("marketplace pricing minimum recommendation must be at least $AGENT_MARKETPLACE_PRICING_MIN_MIN")
-
-            if (pricing.recommendations.min > AGENT_MARKETPLACE_PRICING_MIN_MAX)
-                throw RegistryException("marketplace pricing minimum recommendation must be at most $AGENT_MARKETPLACE_PRICING_MIN_MAX")
-
-            if (pricing.recommendations.max <= pricing.recommendations.min)
-                throw RegistryException("marketplace pricing maximum recommendation must be greater than minimum recommendation")
-        }
-
-        val erc8004 = marketplace.identities?.erc8004
-        if (erc8004 != null) {
-            try {
-                val bytes = Base58.decode(erc8004.wallet)
-                if (bytes.size !in 25..32)
-                    throw RegistryException("marketplace.identities.erc8004.wallet must be between 25 and 32 bytes long, was ${bytes.size}")
-            } catch (e: AddressFormatException) {
-                throw RegistryException("marketplace.identities.erc8004.wallet is not a valid Base58-encoded wallet address: ${e.message}")
-            }
-
-            if (erc8004.endpoints.size > AGENT_MARKETPLACE_ERC8004_ENDPOINTS_MAX_ENTRIES)
-                throw RegistryException("marketplace.identities.erc8004.endpoints cannot exceed $AGENT_MARKETPLACE_ERC8004_ENDPOINTS_MAX_ENTRIES, found ${erc8004.endpoints.size} defined")
-
-            for ((index, endpoint) in erc8004.endpoints.withIndex()) {
-                validateStringLength(
-                    "marketplace.identities.erc8004.endpoints[$index].name",
-                    endpoint.name,
-                    AGENT_MARKETPLACE_ERC8004_ENDPOINTS_NAME_LENGTH
-                )
-
-                if (!endpoint.name.matches(AGENT_MARKETPLACE_ERC8004_ENDPOINTS_NAME_PATTERN))
-                    throw RegistryException("marketplace.identities.erc8004.endpoints[$index].name is not valid.  Marketplace endpoint names must start with an alphabetic character and contain only alphanumeric characters or, underscores or '-'s ")
-
-                validateUri(
-                    "marketplace.identities.erc8004.endpoints[$index].endpoint",
-                    endpoint.endpoint,
-                    AGENT_MARKETPLACE_ERC8004_ENDPOINTS_ENDPOINT_LENGTH,
-                    "https"
-                )
-            }
-        }
-    }
-
-    /**
-     * Validates values in this registry agent to ensure they are compliant with the requirements for the marketplace.
-     *
-     * @throws RegistryException if this registry agent contains any number of invalid values
-     */
-    fun validate() {
-        validateName()
-        validateVersion()
-        validateOptionalAgentInfo()
-        validateRuntimes()
-        validateOptions()
-        validateMarketplace()
-    }
 }
 
 @Serializable
